# swaggers/contents/swagger.yml
openapi: "3.0.2"
info:
    title: Contents API
    version: "0.1"
    description: 영상/교재/위키피디아/아티클 API
servers:
  - url: http://192.168.153.173:8080/
    description: example server address
...
paths:       
...
  /users/{user_id}/books/{book_id}:
    get:
      tags:
        - Book
      summary : 유저의 교재 조회
      description : 유저의 북마크 정보를 포함한 교재 조회
      parameters:
      - name: user_id
        in: path
        description: 사용자 아이디
        required: true
        schema:
          type: string
          format: string
      - name: book_id
        in: path
        description: 교재 아이디
        required: true
        schema:
          type: integer
          format: integer
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                properties:
                  bookId:
                    type: integer
                  bookSrc:
                    type: string
                  title: 
                    type: string
                  createDate: 
                    type: string
                  creatorId: 
                    type: string
                  imgSrc: 
                    type: string
                  description: 
                    type: string
                  bookmark: 
                    type: boolean
  /users/{user_id}/books:
    get:
      tags:
        - Book
      summary : 유저의 교재 리스트 조회
      description : 유저의 북마크 정보를 포함한 교재 리스트 조회
      parameters:
      - name: user_id
        in: path
        description: 사용자 아이디
        required: true
        schema:
          type: string
          format: string
      - name: keyword
        in: query
        description: 검색 키워드
        required: false
        schema:
          type: string
          format: string
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                properties:
                  size:
                    type: integer
                  books:
                    type: array
                    items:
                      type: object
                      properties:
                        bookId:
                          type: integer
                        bookSrc:
                          type: string
                        title: 
                          type: string
                        createDate: 
                          type: string
                        creatorId: 
                          type: string
                        imgSrc: 
                          type: string
                        description: 
                          type: string
                        bookmark: 
                          type: boolean
  /users/{user_id}/books/bookmark:
    get:
      tags:
        - Book
      summary : 유저의 북마크 교재 리스트 조회
      description : 유저가 북마크 선택한 교재 리스트 조회
      parameters:
      - name: user_id
        in: path
        description: 사용자 아이디
        required: true
        schema:
          type: string
          format: string
      - name: keyword
        in: query
        description: 검색 키워드
        required: false
        schema:
          type: string
          format: string
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                properties:
                  size:
                    type: integer
                  books:
                    type: array
                    items:
                      type: object
                      properties:
                        bookId:
                          type: integer
                        bookSrc:
                          type: string
                        title: 
                          type: string
                        createDate: 
                          type: string
                        creatorId: 
                          type: string
                        imgSrc: 
                          type: string
                        description: 
                          type: string
                        bookmark: 
                          type: boolean
  /users/{user_id}/books/{book_id}/bookmark:
    put:
      tags:
        - Book
      summary : 유저의 교재 북마크 추가
      description : 유저가 교재에 북마크 추가
      parameters:
      - name: user_id
        in: path
        description: 사용자 아이디
        required: true
        schema:
          type: string
          format: string
      - name: book_id
        in: path
        description: 교재 아이디
        required: true
        schema:
          type: integer
          format: integer
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                example:
                  isSuccess: true
    delete:
      tags:
        - Book
      summary : 유저의 교재 북마크 삭제
      description : 유저가 교재에 북마크 삭제
      parameters:
      - name: user_id
        in: path
        description: 사용자 아이디
        required: true
        schema:
          type: string
          format: string
      - name: book_id
        in: path
        description: 교재 아이디
        required: true
        schema:
          type: string
          format: string
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                example:
                  isSuccess: true
  /users/{user_id}/books/{book_id}/hit:
    post:
      tags:
      - Book
      summary: 교재 조회수 업데이트
      description: 해당 교재의 조회수를 증가시키고 LRS 저장
      parameters:
      - name: user_id
        in: path
        description: 유저아이디
        required: true
        schema:
          type: string
          format: string
      - name: book_id
        in: path
        description: 교재아이디
        required: true
        schema:
          type: integer
          format: integer
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                example:
                  isSuccess: true                  
  /users/{user_id}/books/{book_id}/quit:
    post:
      tags:
      - Book
      summary: 교재 quit 업데이트
      description: 해당 교재의 quit LRS 저장
      parameters:
      - name: user_id
        in: path
        description: 유저아이디
        required: true
        schema:
          type: string
          format: string
      - name: book_id
        in: path
        description: 교재 아이디
        required: true
        schema:
          type: integer
          format: integer
      - name: duration
        in: query
        description: 컨텐츠 소비 시간
        required: true
        schema:
          type: integer
          format: integer
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                example:
                  isSuccess: true
...
  /users/{user_id}/articles/{article_id}:
    get:
      tags:
        - Article
      summary : 유저의 아티클 조회
      description : 유저의 북마크 정보를 포함한 아티클 조회
      parameters:
      - name: user_id
        in: path
        description: 사용자 아이디
        required: true
        schema:
          type: string
          format: string
      - name: article_id
        in: path
        description: 아티클 아이디
        required: true
        schema:
          type: integer
          format: integer
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                properties:
                  articleId:
                    type: integer
                  articleUrl: 
                    type: string
                  title: 
                    type: string
                  createDate: 
                    type: string
                  creatorId: 
                    type: string
                  description: 
                    type: string
                  imgSrc: 
                    type: string
                  source: 
                    type: string
                  bookmark: 
                    type: boolean
                  uks:
                    type: array
                    items:
                      type: string
  /users/{user_id}/articles:
    get:
      tags:
        - Article
      summary : 유저의 아티클 리스트 조회
      description : 유저의 북마크 정보를 포함한 아티클 리스트 조회
      parameters:
      - name: user_id
        in: path
        description: 사용자 아이디
        required: true
        schema:
          type: string
          format: string
      - name: keyword
        in: query
        description: 검색 키워드
        required: false
        schema:
          type: string
          format: string
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                properties:
                  size:
                    type: integer
                  articles:
                    type: array
                    items:
                      type: object
                      properties:
                        articleId:
                          type: integer
                        articleUrl: 
                          type: string
                        title: 
                          type: string
                        createDate: 
                          type: string
                        creatorId: 
                          type: string
                        description: 
                          type: string
                        imgSrc: 
                          type: string
                        source: 
                          type: string
                        bookmark: 
                          type: boolean
                        uks:
                          type: array
                          items:
                            type: string
  /users/{user_id}/articles/bookmark:
    get:
      tags:
        - Article
      summary : 유저의 북마크 아티클 리스트 조회
      description : 유저가 선택한 북마크 아티클 리스트 조회
      parameters:
      - name: user_id
        in: path
        description: 사용자 아이디
        required: true
        schema:
          type: string
          format: string
      - name: keyword
        in: query
        description: 검색 키워드
        required: false
        schema:
          type: string
          format: string
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                properties:
                  size:
                    type: integer
                  articles:
                    type: array
                    items:
                      type: object
                      properties:
                        articleId:
                          type: integer
                        articleUrl: 
                          type: string
                        title: 
                          type: string
                        createDate: 
                          type: string
                        creatorId: 
                          type: string
                        description: 
                          type: string
                        imgSrc: 
                          type: string
                        source: 
                          type: string
                        bookmark: 
                          type: boolean
                        uks:
                          type: array
                          items:
                            type: string
  /users/{user_id}/articles/{article_id}/bookmark:
    put:
      tags:
        - Article
      summary : 유저의 아티클 북마크 추가
      description : 유저가 아티클에 북마크 추가
      parameters:
      - name: user_id
        in: path
        description: 사용자 아이디
        required: true
        schema:
          type: string
          format: string
      - name: article_id
        in: path
        description: 아티클 아이디
        required: true
        schema:
          type: string
          format: string
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                example:
                  isSuccess: true
    delete:
      tags:
        - Article
      summary : 유저의 아티클 북마크 삭제
      description : 유저가 아티클에 북마크 삭제
      parameters:
      - name: user_id
        in: path
        description: 사용자 아이디
        required: true
        schema:
          type: string
          format: string
      - name: article_id
        in: path
        description: 아티클 아이디
        required: true
        schema:
          type: string
          format: string
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                example:
                  isSuccess: true                    
  /users/{user_id}/articles/{article_id}/hit:
    post:
      tags:
      - Article
      summary: 아티클 조회수 업데이트
      description: 해당 아티클의 조회수를 증가시키고 LRS 저장
      parameters:
      - name: user_id
        in: path
        description: 유저아이디
        required: true
        schema:
          type: string
          format: string
      - name: article_id
        in: path
        description: 아티클아이디
        required: true
        schema:
          type: integer
          format: integer
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                example:
                  isSuccess: true                     
  /users/{user_id}/articles/{article_id}/quit:
    post:
      tags:
      - Article
      summary: 아티클 quit 업데이트
      description: 해당 아티클의 quit LRS 저장
      parameters:
      - name: user_id
        in: path
        description: 유저아이디
        required: true
        schema:
          type: string
          format: string
      - name: article_id
        in: path
        description: 아티클아이디
        required: true
        schema:
          type: integer
          format: integer
      - name: duration
        in: query
        description: 컨텐츠 소비 시간
        required: true
        schema:
          type: integer
          format: integer
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                example:
                  isSuccess: true
...